Fixed floating point issue with json_encode

+++ components/Cart.php
@@ -212,7 +212,7 @@ class Cart extends MallComponent
         return [
             'id'       => $item->prefixedId,
             'name'     => $product->name,
-            'price'    => $item->price()->float,
+            'price'    => $item->price()->decimal,
             'brand'    => optional($item->brand)->name,
             'category' => $item->categories->first()->name,
             'variant'  => optional($variant)->name,
